android: アスキーマップ文字化けの修正、日本語入力（文字エンコーディング）の改善、およびビルドドキュメントの追加
- winandroid.cでのttychar送信時にCP437からUnicodeへの変換テーブルを適用し、アスキーマップの文字化けを防ぐように改善
- winandroid.cの各入力受付（getline, askname等）にてGetStringUTFCharsを用いてUTF-8での日本語入力を適切に処理できるように改善
- debuglog関数についてスレッド安全な__android_log_printを使用したネイティブログ出力への刷新
- Android向けビルド手順書(README_JP.md)および自動ビルドスクリプト(build_android.ps1)の追加

+++ sys/android/winandroid.c
@@ -1,4 +1,5 @@
-/* Modified by NetHackJP contributor @satokiyon; latest change date: 2026-06-26. */
+/* Modified by NetHackJP contributor @satokiyon; latest change date: 2026-06-27. */
+#include <android/log.h>
 #include <string.h>
 #include <errno.h>
 #include <jni.h>
@@ -357,9 +358,14 @@ void debuglog(const char *fmt, ...)
 		strcpy(buf, "(null)");
 	}
 
-	jbyteArray jstr = create_bytearray(buf);
-	JNICallV(jDebugLog, jstr);
-	destroy_jobject(jstr);
+	__android_log_print(ANDROID_LOG_DEBUG, "NetHackJP", "%s", buf);
+
+	if (jEnv && jAppInstance && jDebugLog)
+	{
+		jbyteArray jstr = create_bytearray(buf);
+		JNICallV(jDebugLog, jstr);
+		destroy_jobject(jstr);
+	}
 }
 
 //____________________________________________________________________________________
@@ -1384,6 +1390,26 @@ void and_update_positionbar(char *features)
 
 #endif
 
+/* NetHackJP: useCP437Decoder = falseの際、アスキーマップの文字化けを防ぐためのCP437からUnicodeへの変換テーブル */
+static const unsigned short cp437_to_unicode[256] = {
+    0x00A0, 0x263A, 0x263B, 0x2665, 0x2666, 0x2663, 0x2660, 0x2022, 0x25D8, 0x25CB, 0x25D9, 0x2660, 0x2661, 0x266A, 0x266B, 0x2609,
+    0x25BA, 0x25C4, 0x2195, 0x203C, 0x00B6, 0x00A7, 0x25AC, 0x2607, 0x2191, 0x2193, 0x2192, 0x2190, 0x221F, 0x2194, 0x25B2, 0x25BC,
+    0x0020, 0x0021, 0x0022, 0x0023, 0x0024, 0x0025, 0x0026, 0x0027, 0x0028, 0x0029, 0x002A, 0x002B, 0x002C, 0x002D, 0x002E, 0x002F,
+    0x0030, 0x0031, 0x0032, 0x0033, 0x0034, 0x0035, 0x0036, 0x0037, 0x0038, 0x0039, 0x003A, 0x003B, 0x003C, 0x003D, 0x003E, 0x003F,
+    0x0040, 0x0041, 0x0042, 0x0043, 0x0044, 0x0045, 0x0046, 0x0047, 0x0048, 0x0049, 0x004A, 0x004B, 0x004C, 0x004D, 0x004E, 0x004F,
+    0x0050, 0x0051, 0x0052, 0x0053, 0x0054, 0x0055, 0x0056, 0x0057, 0x0058, 0x0059, 0x005A, 0x005B, 0x005C, 0x005D, 0x005E, 0x005F,
+    0x0060, 0x0061, 0x0062, 0x0063, 0x0064, 0x0065, 0x0066, 0x0067, 0x0068, 0x0069, 0x006A, 0x006B, 0x006C, 0x006D, 0x006E, 0x006F,
+    0x0070, 0x0071, 0x0072, 0x0073, 0x0074, 0x0075, 0x0076, 0x0077, 0x0078, 0x0079, 0x007A, 0x007B, 0x007C, 0x007D, 0x007E, 0x2206,
+    0x00C7, 0x00FC, 0x00E9, 0x00E2, 0x00E4, 0x00E0, 0x00E5, 0x00E7, 0x00EA, 0x00EB, 0x00E8, 0x00EF, 0x00EE, 0x00EC, 0x00C4, 0x00C5,
+    0x00C9, 0x00E6, 0x00C6, 0x00F4, 0x00F6, 0x00F2, 0x00FB, 0x00F9, 0x00FF, 0x00D6, 0x00DC, 0x00A2, 0x00A3, 0x00A5, 0x20A3, 0x0192,
+    0x00E1, 0x00ED, 0x00F3, 0x00FA, 0x00F1, 0x00D1, 0x00AA, 0x00BA, 0x00BF, 0x2310, 0x00AC, 0x00BD, 0x00BC, 0x00A1, 0x00AB, 0x00BB,
+    0x2591, 0x2592, 0x2591, 0x2502, 0x2524, 0x2561, 0x2562, 0x2556, 0x2555, 0x2563, 0x2551, 0x2557, 0x255D, 0x255C, 0x255B, 0x2510,
+    0x2514, 0x2534, 0x252C, 0x251C, 0x2500, 0x253C, 0x255E, 0x255F, 0x255A, 0x2554, 0x2569, 0x2566, 0x2560, 0x2550, 0x256C, 0x2567,
+    0x2568, 0x2564, 0x2565, 0x2559, 0x2558, 0x2552, 0x2553, 0x256B, 0x256A, 0x2518, 0x250C, 0x2588, 0x2584, 0x258C, 0x2590, 0x2580,
+    0x03B1, 0x00DF, 0x0393, 0x03C0, 0x03A3, 0x03C3, 0x00B5, 0x03C4, 0x03A6, 0x0398, 0x03A9, 0x03B4, 0x221E, 0x03C6, 0x25A0, 0x002B,
+    0x2261, 0x00B1, 0x2265, 0x2264, 0x2320, 0x2321, 0x00F7, 0x2248, 0x00B0, 0x2219, 0x00B7, 0x221A, 0x207F, 0x00B2, 0x25A0, 0x00A0
+};
+
 //____________________________________________________________________________________
 //print_glyph(window, x, y, glyph)
 //		-- Print the glyph at (x,y) on the given window.  Glyphs are
@@ -1406,8 +1432,16 @@ void and_print_glyph(winid wid, coordxy x, coordxy y, const glyph_info *glyphinf
         special &= ~MG_OBJPILE;
     if(!iflags.use_inverse)
         special &= ~MG_DETECT;
-	JNICallV(jPrintTile, wid, x, y, tile, glyphinfo->ttychar, nhcolor_to_RGB(glyphinfo->gm.sym.color), special);
+
+    /* NetHackJP: CP437からUnicodeへの変換を適用 */
+    int ch = glyphinfo->ttychar;
+    if (ch >= 0 && ch < 256) {
+        ch = cp437_to_unicode[ch];
+    }
+
+	JNICallV(jPrintTile, wid, x, y, tile, ch, nhcolor_to_RGB(glyphinfo->gm.sym.color), special);
 }
+
 
 //____________________________________________________________________________________
 // raw_print(str)	-- Print directly to a screen, or otherwise guarantee that
@@ -1806,46 +1840,42 @@ void and_n_getline_r(const char* question, char* buf, int nMax, int showLog, int
 	n = (*jEnv)->GetStringLength(jEnv, jstr);
 	if(n >= nMax)
 		n = nMax - 1;
-    i = 0;
+
 	if(n > 0)
 	{
-		pChars = (*jEnv)->GetStringChars(jEnv, jstr, 0);
-	//debuglog("    returned %c %s", *pChars, pChars);
-		if(*pChars == 0x80)
-		{
-			// special case: ABORT
-			if(!program_state.gameover && program_state.something_worth_saving)
+		const char *utf8 = (*jEnv)->GetStringUTFChars(jEnv, jstr, 0);
+		if (utf8) {
+			if(utf8[0] == 0x80)
 			{
-				buf[0] = '\033';
-				i = 1;
+				// special case: ABORT
+				if(!program_state.gameover && program_state.something_worth_saving)
+				{
+					buf[0] = '\033';
+					buf[1] = '\0';
+				}
+				else
+				{
+					(*jEnv)->ReleaseStringUTFChars(jEnv, jstr, utf8);
+					destroy_jobject(jstr);
+					and_n_getline_r(question, buf, nMax, showLog, 1);
+					return;
+				}
 			}
 			else
 			{
-				(*jEnv)->ReleaseStringChars(jEnv, jstr, pChars);
-				destroy_jobject(jstr);
-				and_n_getline_r(question, buf, nMax, showLog, 1);
-				return;
+				strncpy(buf, utf8, nMax-1);
+				buf[nMax-1] = '\0';
 			}
+			(*jEnv)->ReleaseStringUTFChars(jEnv, jstr, utf8);
+		} else {
+			buf[0] = '\0';
 		}
-		else if(*pChars == '\033')
-		{
-			buf[0] = '\033';
-			i = 1;
-		}
-		else
-		{
-			for(; i < n; i++)
-			{
-				if(isprint(pChars[i]))
-					buf[i] = pChars[i];
-				else
-					buf[i] = '?';
-			}
-		}
-		(*jEnv)->ReleaseStringChars(jEnv, jstr, pChars);
+	}
+	else
+	{
+		buf[0] = '\0';
 	}
 	destroy_jobject(jstr);
-	buf[i] = 0;
 }
 
 //____________________________________________________________________________________
@@ -1878,7 +1908,6 @@ void and_askname()
 //	debuglog("ask name");
 
 	int i, n, w;
-	const jchar* pChars;
 	jstring jstr;
 
 	char** saves = get_saved_games();
@@ -1901,34 +1930,33 @@ void and_askname()
     	destroy_jobject((*jEnv)->GetObjectArrayElement(jEnv, strings, i));
 	destroy_jobject(strings);
 
-	n = (*jEnv)->GetStringLength(jEnv, jstr) - 1;
-	w = n;
-	if(n >= PL_NSIZ)
-		n = PL_NSIZ - 1;
-    i = 0;
-	if(n > 0)
-	{
-		pChars = (*jEnv)->GetStringChars(jEnv, jstr, 0);
-		if(*pChars == 0x80 || *pChars == '\033')
-		{
+	const char *utf8 = (*jEnv)->GetStringUTFChars(jEnv, jstr, 0);
+	if (utf8) {
+		if (utf8[0] == 0x80 || utf8[0] == '\033') {
 			clearlocks();
 			and_exit_nhwindows("bye");
 			nh_terminate(EXIT_SUCCESS);
 		}
 
-		if( pChars[w] == '1' )
-			wizard = TRUE;
+		int len = (int) strlen(utf8);
+		if (len > 1) {
+			if (utf8[len - 1] == '1')
+				wizard = TRUE;
 
-		for(; i < n; i++)
-		{
-			if(isprint(pChars[i]))
-				svp.plname[i] = pChars[i];
-			else
-				svp.plname[i] = '?';
+			int name_len = len - 1;
+			if (name_len >= PL_NSIZ)
+				name_len = PL_NSIZ - 1;
+
+			strncpy(svp.plname, utf8, name_len);
+			svp.plname[name_len] = '\0';
+			utf8_truncate(svp.plname, PL_NSIZ - 1);
+		} else {
+			svp.plname[0] = '\0';
 		}
-		(*jEnv)->ReleaseStringChars(jEnv, jstr, pChars);
+		(*jEnv)->ReleaseStringUTFChars(jEnv, jstr, utf8);
+	} else {
+		svp.plname[0] = '\0';
 	}
-	svp.plname[i] = 0;
 	destroy_jobject(jstr);
 }
 
@@ -2261,12 +2289,18 @@ void and_get_dumplog_dir(char* buf)
 
 	if(n > 0 && n < BUFSZ - 1)
 	{
-		pChars = (*jEnv)->GetStringChars(jEnv, jstr, 0);
-		for(i = 0; i < n; i++)
-			buf[i] = pChars[i];
-		(*jEnv)->ReleaseStringChars(jEnv, jstr, pChars);
-		if(buf[n - 1] != '/')
-			buf[n++] = '/';
+		const char *utf8 = (*jEnv)->GetStringUTFChars(jEnv, jstr, 0);
+		if (utf8) {
+			strncpy(buf, utf8, n);
+			buf[n] = '\0';
+			(*jEnv)->ReleaseStringUTFChars(jEnv, jstr, utf8);
+			if (buf[n - 1] != '/') {
+				buf[n++] = '/';
+				buf[n] = '\0';
+			}
+		} else {
+			buf[0] = '\0';
+		}
 	}
 	else
 		n = 0;
